refactor: dao flow refactor
