+++ src/main/java/org/jepria/compat/server/upload/blob/BinaryFileUploadImpl.java
@@ -1,12 +1,10 @@
 package org.jepria.compat.server.upload.blob;
 
 import org.jepria.compat.server.db.blob.BinaryLargeObject;
-import org.jepria.compat.server.dao.CallContext;
 import org.jepria.compat.server.exceptions.SpaceException;
 import org.jepria.compat.server.upload.AbstractFileUpload;
 import org.jepria.compat.shared.exceptions.ApplicationException;
 
-import java.util.List;
 import java.util.Map;
 
 /**
@@ -38,8 +36,6 @@ public class BinaryFileUploadImpl extends AbstractFileUpload implements BinaryFi
     } catch (ApplicationException ex) {
       cancel();
       throw ex;
-    } finally {
-      storedContext = CallContext.detach();
     }
 
     return result;
@@ -55,19 +51,20 @@ public class BinaryFileUploadImpl extends AbstractFileUpload implements BinaryFi
     } catch (ApplicationException ex) {
       cancel();
       throw ex;
-    } finally {
-      storedContext = CallContext.detach();
     }
-
     return result;
   }
-
+  
+  @Override
+  public int beginWrite(Object rowId) {
+    throw new UnsupportedOperationException();
+  }
+  
   /**
    * {@inheritDoc}
    */
   @Override
   public void continueWrite(byte[] dataBlock) throws SpaceException {
-    CallContext.attach(storedContext);
     boolean cancelled = false;
     try {
       ((BinaryLargeObject) super.largeObject).continueWrite(dataBlock);
@@ -88,7 +85,6 @@ public class BinaryFileUploadImpl extends AbstractFileUpload implements BinaryFi
           e.printStackTrace();
         }
       }
-      storedContext = CallContext.detach();
     }
   }
 }
