+++ src/main/java/org/jepria/server/data/dao/Dao.java
@@ -1,4 +1,4 @@
-package org.jepria.server.data;
+package org.jepria.server.data.dao;
 
 import java.util.List;
 import java.util.Map;
