+++ src/main/java/org/jepria/server/transaction/TransactionFactory.java
@@ -1,13 +1,13 @@
-package org.jepria.compat.server.dao.transaction;
+package org.jepria.server.transaction;
 
-import org.jepria.compat.server.dao.transaction.annotation.After;
-import org.jepria.compat.server.dao.transaction.annotation.Before;
-import org.jepria.compat.server.dao.transaction.handler.EndTransactionHandler;
-import org.jepria.compat.server.dao.transaction.handler.EndTransactionHandlerImpl;
-import org.jepria.compat.server.dao.transaction.handler.StartTransactionHandler;
-import org.jepria.compat.server.dao.transaction.handler.StartTransactionHandlerImpl;
-import org.jepria.compat.server.db.Db;
 import org.apache.log4j.Logger;
+import org.jepria.server.transaction.annotation.After;
+import org.jepria.server.transaction.annotation.Before;
+import org.jepria.server.transaction.annotation.Transaction;
+import org.jepria.server.transaction.handler.EndTransactionHandler;
+import org.jepria.server.transaction.handler.EndTransactionHandlerImpl;
+import org.jepria.server.transaction.handler.StartTransactionHandler;
+import org.jepria.server.transaction.handler.StartTransactionHandlerImpl;
 
 import java.lang.reflect.InvocationHandler;
 import java.lang.reflect.Method;
@@ -17,9 +17,9 @@ import java.lang.reflect.Proxy;
  * Фабрика, создающая прокси для выполнения методов Dao в рамках одной транзакции.
  */
 public class TransactionFactory {
-
+  
   protected static Logger logger = Logger.getLogger(TransactionFactory.class.getName());
-
+  
   /**
    * Обработчик вызова метода Dao, обеспечивающий его выполнение в рамках одной транзакции.<br/>
    * Механизм работы следующий:
@@ -29,10 +29,11 @@ public class TransactionFactory {
    *   оно перехватывается.</li>
    *   <li>Вызывается обработчик завершения транзакции.</li>
    * </ul>
+   *
    * @param <D> интерфейс Dao
    */
   private static class TransactionInvocationHandler<D> implements InvocationHandler {
-
+    
     /**
      * Объект Dao.
      */
@@ -45,67 +46,72 @@ public class TransactionFactory {
      * Имя модуля для передачи в DB.
      */
     private final String moduleName;
-
+    
     /**
      * Создаёт экземпляр транзакционного обработчика.
-     * @param dao объект Dao
+     *
+     * @param dao                объект Dao
      * @param dataSourceJndiName JNDI-имя источника данных
-     * @param moduleName имя модуля для передачи в DB
+     * @param moduleName         имя модуля для передачи в DB
      */
     public TransactionInvocationHandler(D dao, String dataSourceJndiName, String moduleName) {
       this.dao = dao;
       this.dataSourceJndiName = dataSourceJndiName;
       this.moduleName = moduleName;
     }
-
+    
     /**
      * {@inheritDoc}
      */
     @Override
     public Object invoke(Object proxy, Method method, Object[] args) throws Throwable {
       Class<?> daoClass = dao.getClass();
-      Method implementingMethod = daoClass.getMethod(
-          method.getName(), method.getParameterTypes());
-      
-      Before before = implementingMethod.getAnnotation(Before.class);
-      Class<? extends StartTransactionHandler> startTransactionHandlerClass =
-          before != null ? before.startTransactionHandler() : StartTransactionHandlerImpl.class;
+      if (dao.getClass().getDeclaredMethod(method.getName(), method.getParameterTypes()).isAnnotationPresent(Transaction.class)
+          || daoClass.isAnnotationPresent(Transaction.class)) {
+        Method implementingMethod = daoClass.getMethod(
+            method.getName(), method.getParameterTypes());
         
-      After after = implementingMethod.getAnnotation(After.class);
-      Class<? extends EndTransactionHandler> endTransactionHandlerClass =
-          after != null ? after.endTransactionHandler() : EndTransactionHandlerImpl.class;
-          
-      Db db = startTransactionHandlerClass.newInstance().handle(dataSourceJndiName, moduleName);
-      Throwable caught = null;
-      Object result = null;
-      synchronized (db) {
+        Before before = implementingMethod.getAnnotation(Before.class);
+        Class<? extends StartTransactionHandler> startTransactionHandlerClass =
+            before != null ? before.startTransactionHandler() : StartTransactionHandlerImpl.class;
+        
+        After after = implementingMethod.getAnnotation(After.class);
+        Class<? extends EndTransactionHandler> endTransactionHandlerClass =
+            after != null ? after.endTransactionHandler() : EndTransactionHandlerImpl.class;
+        
+        startTransactionHandlerClass.newInstance().handle(dataSourceJndiName, moduleName);
+        Throwable caught = null;
+        Object result = null;
         try {
           final long startTime = System.currentTimeMillis();
           result = method.invoke(dao, args);
-          logger.trace(dao.getClass() +"." + method.getName() + " execution time: " + (System.currentTimeMillis() - startTime)/1000.00 + " (seconds)");
-        } catch(Exception exc) {
+          logger.trace(dao.getClass() + "." + method.getName() + " execution time: " + (System.currentTimeMillis() - startTime) / 1000.00 + " (seconds)");
+        } catch (Exception exc) {
           /*
            * Необходимо вызвать getCause(), поскольку выброшенное из Dao исключение
            * будет обёрнуто в InvocationTargetException.
            */
           caught = exc.getCause();
         }
-
+        
         endTransactionHandlerClass.newInstance().handle(caught);
-
+        
         if (caught != null) {
           throw caught;
         }
+        return result;
+      } else {
+        return method.invoke(dao, args);
       }
-      return result;
     }
   }
-
+  
   /**
    * Создаёт прокси для переданного Dao.
-   * @param dao объект Dao
+   *
+   * @param dao                объект Dao
    * @param dataSourceJndiName JNDI-имя источника данных
-   * @param moduleName имя модуля
+   * @param moduleName         имя модуля
    * @return созданный прокси
    */
   @SuppressWarnings("unchecked")
