+++ src/main/java/org/jepria/server/transaction/annotation/Transaction.java
@@ -1,0 +1,11 @@
+package org.jepria.server.transaction.annotation;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Target({ElementType.METHOD, ElementType.TYPE})
+@Retention(RetentionPolicy.RUNTIME)
+public @interface Transaction {
+}
