+++ src/main/java/org/jepria/server/transaction/handler/EndTransactionHandlerImpl.java
@@ -1,6 +1,6 @@
-package org.jepria.compat.server.dao.transaction.handler;
+package org.jepria.server.transaction.handler;
 
-import org.jepria.compat.server.dao.CallContext;
+import org.jepria.server.data.sql.ConnectionContext;
 
 import java.sql.SQLException;
 
@@ -15,10 +15,10 @@ public class EndTransactionHandlerImpl implements EndTransactionHandler {
    * Выполняет следующие действия:
    * <ul>
    *   <li>Если в ходе транзакции не возникло исключения, то транзакция фиксируется
-   *   ({@link CallContext#commit()}).</li>
+   *   ({@link ConnectionContext#commit()}).</li>
    *   <li>Если было перехвачено исключение (<code>caught != null</code>), транзакция
-   *   откатывается ({@link CallContext#rollback()}).</li>
-   *   <li>С помощью {@link CallContext#end()} освобождаются ресурсы.
+   *   откатывается ({@link ConnectionContext#rollback()}).</li>
+   *   <li>С помощью {@link ConnectionContext#close()} ()} освобождаются ресурсы.
    *   <li>Если в ходе транзакции возникло исключение, или же оно возникло во время
    *   commit либо rollback, выбрасывается последнее возникшее исключение.</li>
    * </ul>
@@ -28,17 +28,17 @@ public class EndTransactionHandlerImpl implements EndTransactionHandler {
   public void handle(Throwable caught) {
     try {
       if (caught == null) {
-        CallContext.commit();
+        ConnectionContext.getInstance().getConnection().commit();
       }
       else {
-        CallContext.rollback();
+        ConnectionContext.getInstance().getConnection().rollback();
       }
     } catch (SQLException e) {
       // Необходимо сигнализировать о последнем выброшенном исключении.
       caught = e;
     }
     finally {
-      CallContext.end();
+      ConnectionContext.getInstance().end();
     }
   }
